# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Paskihub API Documentation",
        "title": "Paskihub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3010",
    "basePath": "/",
    "paths": {
...
        "/api/v1/eo/events/{eventId}/teams": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of team registrations for an event (Event Organizer)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Team"
                ],
                "summary": "Get list of teams",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event Level ID filter",
                        "name": "event_level_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/eo/events/{eventId}/teams/{registrationId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve details of a specific team registration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Team"
                ],
                "summary": "Get team details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registration ID",
                        "name": "registrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/eo/events/{eventId}/teams/{registrationId}/approve": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Approve a team's registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Team"
                ],
                "summary": "Approve team registration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registration ID",
                        "name": "registrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Approval Details",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EOTeamApproveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/eo/events/{eventId}/teams/{registrationId}/kick": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Disqualify or kick an approved team",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Team"
                ],
                "summary": "Kick team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registration ID",
                        "name": "registrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/eo/events/{eventId}/teams/{registrationId}/reject": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Reject a pending team registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EO Team"
                ],
                "summary": "Reject team registration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registration ID",
                        "name": "registrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection Reason",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EOTeamRejectReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/peserta/teams": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of teams for the current participant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant Team"
                ],
                "summary": "Get participant teams",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Upload team details including logo, recommendation letter, and members",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant Team"
                ],
                "summary": "Create a new participant team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coach Name",
                        "name": "pelatih_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Team Logo",
                        "name": "logo_team",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Recommendation Letter",
                        "name": "surat_rekomendasi",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/peserta/teams/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve details of a specific team",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant Team"
                ],
                "summary": "Get team details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Update team details including logo, recommendation letter, and members",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant Team"
                ],
                "summary": "Update participant team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Team Name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Coach Name",
                        "name": "pelatih_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Team Logo",
                        "name": "logo_team",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Recommendation Letter",
                        "name": "surat_rekomendasi",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a team by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant Team"
                ],
                "summary": "Delete participant team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.EOTeamApproveReq": {
            "type": "object",
            "required": [
                "payment_status"
            ],
            "properties": {
                "payment_status": {
                    "enum": [
                        "DP_PAID",
                        "FULL_PAID"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.RegistrationStatus"
                        }
                    ]
                }
            }
        },
        "dto.EOTeamRejectReq": {
            "type": "object",
            "required": [
                "rejection_reason"
            ],
            "properties": {
                "rejection_reason": {
                    "type": "string"
                }
            }
        },
...
        "enums.RegistrationStatus": {
            "type": "string",
            "enum": [
                "WAITING",
                "DP_PAID",
                "FULL_PAID",
                "REJECTED"
            ],
            "x-enum-varnames": [
                "Waiting",
                "DPPaid",
                "FullPaid",
                "Rejected"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for all endpoints. Format: Key {api_key}. Example: Key abc123",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token. Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}